+++ src/lschvkbb.f90
@@ -147,7 +147,7 @@ subroutine lschvkbb(nn,ll,nvkb,ierr,ee,emin,emax, &
 ! start inward integration at 10*classical turning
 ! point with simple exponential
 
-         nin=mch+2.3d0/al
+         nin=int(mch+2.3d0/al)
          if(nin+4>mmax) nin=mmax-4
          xkap=dsqrt(sls/rr(nin)**2 + 2.0d0*(vloc(nin)-ee))
